MS2: implemented nextauth and refacored auth code
