+++ src/management-system-v2/app/(dashboard)/processes/page.tsx
@@ -1,11 +1,9 @@
-'use client';
 import { FC } from 'react';
 import Processes from '@/components/processes';
 import Content from '@/components/content';
-import { login } from '@/lib/iam';
-import { Button, Result, Space } from 'antd';
-import { LoginOutlined } from '@ant-design/icons';
-import Auth from '@/lib/AuthCanWrapper';
+import { Result, Space } from 'antd';
+import NotLoggedInFallback from './not-logged-in-fallback';
+import Auth from '@/lib/serverAuthComponents';
 
 const ProcessesPage: FC = () => {
   return (
@@ -24,18 +22,7 @@ export default Auth(
     fallback: (
       <Result status="403" title="Not allowed" subTitle="You're not allowed to view processes" />
     ),
-    notLoggedIn: (
-      <Result
-        status="403"
-        title="You're not logged in"
-        subTitle="Sorry, you have to be logged in to use the app"
-        extra={
-          <Button type="primary" icon={<LoginOutlined />} onClick={login}>
-            Login
-          </Button>
-        }
-      />
-    ),
+    notLoggedIn: <NotLoggedInFallback />,
   },
   ProcessesPage,
 );
