+++ src/management-system-v2/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,44 @@
+import { AuthOptions } from 'next-auth';
+import NextAuth from 'next-auth/next';
+import Auth0Provider from 'next-auth/providers/auth0';
+import { User } from '@/types/next-auth';
+
+export const nextAuthOptions: AuthOptions = {
+  secret: process.env.NEXTAUTH_SECRET,
+  providers: [
+    Auth0Provider({
+      clientId: process.env.AUTH0_CLIENT_ID as string,
+      clientSecret: process.env.AUTH0_CLIENT_SECRET as string,
+      issuer: process.env.AUTH0_ISSUER,
+      authorization: {
+        params: {
+          scope: process.env.AUTH0_SCOPE,
+        },
+      },
+      profile(profile) {
+        return {
+          id: profile.sub,
+          email: profile.email,
+          image: profile.picture,
+          firstName: profile.given_name,
+          lastName: profile.family_name,
+          username: profile.preferred_username,
+        };
+      },
+    }),
+  ],
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) token.user = user as User;
+      return token;
+    },
+    session({ session, token }) {
+      if (token.user) session.user = token.user;
+      return session;
+    },
+  },
+};
+
+const handler = NextAuth(nextAuthOptions);
+
+export { handler as GET, handler as POST };
