+++ src/management-system-v2/components/layout.tsx
@@ -19,8 +19,9 @@ import {
 import Image from 'next/image';
 import { usePathname, useRouter } from 'next/navigation';
 import cn from 'classnames';
-import { useAuthStore } from '@/lib/iam';
+import { useAbilityStore } from '@/lib/abilityStore';
 import Link from 'next/link';
+import { useSession } from 'next-auth/react';
 
 const items: MenuProps['items'] = [
   {
@@ -54,8 +55,9 @@ const Layout: FC<PropsWithChildren> = ({ children }) => {
   const router = useRouter();
   const activeSegment = usePathname().slice(1) || 'processes';
   const [collapsed, setCollapsed] = useState(false);
-  const ability = useAuthStore((state) => state.ability);
-  const loggedIn = useAuthStore((state) => state.loggedIn);
+  const { status } = useSession();
+  const loggedIn = status === 'authenticated';
+  const ability = useAbilityStore((state) => state.ability);
   const breakpoint = Grid.useBreakpoint();
 
   return (
